rename interp_polynomial to polynomial_interp to match change in interp_mean
